+++ example/lib/main.dart
@@ -2,6 +2,7 @@ import 'package:example/list.dart';
 import 'package:example/row.dart';
 import 'package:example/scrollList.dart';
 import 'package:flutter/material.dart';
+import 'package:left_scroll_actions/cupertinoLeftScroll.dart';
 import 'package:left_scroll_actions/left_scroll_actions.dart';
 
 void main() => runApp(MyApp());
@@ -50,21 +51,21 @@ class _HomePageState extends State<HomePage> {
                 },
               ),
             ),
-            Container(
-              padding: EdgeInsets.all(12),
-              child: MaterialButton(
-                color: Colors.blue,
-                child: Text(
-                  'ClosableListView Usage Demo',
-                  style: TextStyle(color: Colors.white),
-                ),
-                onPressed: () {
-                  Navigator.of(context).push(MaterialPageRoute(
-                    builder: (context) => ClosableListPage(),
-                  ));
-                },
-              ),
-            ),
+            // Container(
+            //   padding: EdgeInsets.all(12),
+            //   child: MaterialButton(
+            //     color: Colors.blue,
+            //     child: Text(
+            //       'ClosableListView Usage Demo',
+            //       style: TextStyle(color: Colors.white),
+            //     ),
+            //     onPressed: () {
+            //       Navigator.of(context).push(MaterialPageRoute(
+            //         builder: (context) => ClosableListPage(),
+            //       ));
+            //     },
+            //   ),
+            // ),
             Container(height: 50),
             Container(
               padding: EdgeInsets.only(top: 12, left: 8, bottom: 8),
@@ -98,6 +99,7 @@ class _HomePageState extends State<HomePage> {
                 print('tap row');
               },
             ),
+
             LeftScroll(
               child: Container(
                 height: 60,
@@ -115,6 +117,69 @@ class _HomePageState extends State<HomePage> {
                   color: Colors.orange,
                 ),
               ],
+            ),
+            Container(
+              padding: EdgeInsets.only(top: 12, left: 8, bottom: 8),
+              child: Text('CupertinoLeftScrollRow'),
+            ),
+            CupertinoLeftScroll(
+              buttonWidth: 80,
+              child: Container(
+                height: 60,
+                color: Colors.white,
+                alignment: Alignment.center,
+                child: Text('👈 Try Scroll Left(iOS style)'),
+              ),
+              buttons: <Widget>[
+                LeftScrollItem(
+                  text: 'delete',
+                  color: Colors.red,
+                  onTap: () {
+                    print('delete');
+                  },
+                ),
+                LeftScrollItem(
+                  text: 'Edit',
+                  color: Colors.orange,
+                  onTap: () {
+                    print('edit');
+                  },
+                ),
+              ],
+              onTap: () {
+                print('tap row');
+              },
+            ),
+            CupertinoLeftScroll(
+              opacityChange: true,
+              buttonWidth: 80,
+              child: Container(
+                height: 60,
+                color: Colors.white,
+                alignment: Alignment.center,
+                child: Text('👈 Try iOS style with opa change'),
+              ),
+              buttons: <Widget>[
+                LeftScrollItem(
+                  text: 'delete',
+                  textColor: Colors.red,
+                  color: Colors.red.withOpacity(0),
+                  onTap: () {
+                    print('delete');
+                  },
+                ),
+                LeftScrollItem(
+                  text: 'Edit',
+                  textColor: Colors.orange,
+                  color: Colors.orange.withOpacity(0),
+                  onTap: () {
+                    print('edit');
+                  },
+                ),
+              ],
+              onTap: () {
+                print('tap row');
+              },
             ),
             Container(height: 50),
             Container(